# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "planaffe",
    "description": "The HTTP surface of one planaffe instance. Every instance is at its own address; everything but GET /version takes `Authorization: Bearer <token>`.",
    "version": "0.0.0-dev"
  },
  "paths": {
...
    "/projects": {
      "post": {
        "tags": [
          "planaffe"
        ],
        "summary": "Create a project with its key and the `kind` label group. Users only.",
        "operationId": "CreateProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "planaffe"
        ],
        "summary": "Every project the caller sees. Not paginated.",
        "operationId": "ListProjects",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/projects/{key}": {
      "get": {
        "tags": [
          "planaffe"
        ],
        "summary": "One project by key.",
        "operationId": "ReadProject",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "planaffe"
        ],
        "summary": "Change the name or the switches. Users only; the key is immutable.",
        "operationId": "ChangeProject",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "planaffe"
        ],
        "summary": "Soft-delete the project with everything in it. Administrators only.",
        "operationId": "DeleteProject",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{key}/restore": {
      "post": {
        "tags": [
          "planaffe"
        ],
        "summary": "Bring a deleted project back, with everything in it. Administrators only.",
        "operationId": "RestoreProject",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{key}/next": {
      "get": {
        "tags": [
          "planaffe"
        ],
        "summary": "What the caller would be handed, in that order — the ready-for-agents list — and why the rest is not on it.",
        "operationId": "PreviewNext",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ready",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "epic",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextPage"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "planaffe"
        ],
        "summary": "Take the highest-ranked workable issue and claim it for the caller, in one transaction. 200 with `issue: null` when nothing is workable.",
        "operationId": "TakeNext",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextAnswer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlockerLink": {
        "required": [
          "key",
          "title",
          "status",
          "open"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IssueStatus"
              }
            ]
          },
          "open": {
            "type": "boolean"
          }
        }
      },
      "BlockerRef": {
        "required": [
          "key",
          "open"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": [
              "null",
              "string"
            ]
          },
          "open": {
            "type": "boolean"
          }
        }
      },
...
      "ChangeProjectRequest": {
        "required": [
          "name",
          "triage_required",
          "review_required"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "triage_required": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "review_required": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "Claim": {
        "required": [
          "holder",
          "since",
          "expires_at"
        ],
        "type": "object",
        "properties": {
          "holder": {
            "$ref": "#/components/schemas/IdentityRef"
          },
          "since": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "Comment": {
        "required": [
          "id",
          "author",
          "body",
          "created_at"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "author": {
            "$ref": "#/components/schemas/IdentityRef"
          },
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateProjectRequest": {
        "required": [
          "key",
          "name",
          "triage_required",
          "review_required"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": [
              "null",
              "string"
            ]
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "triage_required": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "review_required": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
      "EpicRef": {
        "required": [
          "key",
          "title",
          "description",
          "status"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EpicStatus"
          }
        }
      },
      "EpicStatus": {
        "enum": [
          "open",
          "closed"
        ],
        "type": "string"
      },
...
      "IdentityKind": {
        "enum": [
          "user",
          "agent"
        ],
        "type": "string"
      },
      "IdentityRef": {
        "required": [
          "id",
          "kind",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/IdentityKind"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Issue": {
        "required": [
          "key",
          "project",
          "title",
          "description",
          "result",
          "status",
          "ready",
          "priority",
          "labels",
          "epic",
          "assignee",
          "claim",
          "author",
          "blocked_by",
          "blocks",
          "open_questions",
          "open_blockers",
          "open_sub_issues",
          "comments",
          "questions",
          "project_context",
          "created_at",
          "updated_at",
          "closed_at"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "result": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "ready": {
            "type": "boolean"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "labels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Label"
            }
          },
          "epic": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EpicRef"
              }
            ]
          },
          "assignee": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentityRef"
              }
            ]
          },
          "claim": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Claim"
              }
            ]
          },
          "author": {
            "$ref": "#/components/schemas/IdentityRef"
          },
          "blocked_by": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockerLink"
            }
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockerLink"
            }
          },
          "open_questions": {
            "type": "integer",
            "format": "int32"
          },
          "open_blockers": {
            "type": "integer",
            "format": "int32"
          },
          "open_sub_issues": {
            "type": "integer",
            "format": "int32"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          },
          "project_context": {
            "$ref": "#/components/schemas/ProjectContext"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "closed_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "IssueStatus": {
        "enum": [
          "backlog",
          "todo",
          "in_progress",
          "review",
          "done",
          "canceled"
        ],
        "type": "string"
      },
      "IssueSummary": {
        "required": [
          "key",
          "project",
          "title",
          "status",
          "ready",
          "priority",
          "labels",
          "epic",
          "assignee",
          "claim",
          "blocked_by",
          "open_questions",
          "open_blockers",
          "open_sub_issues",
          "created_at",
          "updated_at",
          "closed_at",
          "deleted_at",
          "deleted_by"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "ready": {
            "type": "boolean"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "epic": {
            "type": [
              "null",
              "string"
            ]
          },
          "assignee": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentityRef"
              }
            ]
          },
          "claim": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Claim"
              }
            ]
          },
          "blocked_by": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockerRef"
            }
          },
          "open_questions": {
            "type": "integer",
            "format": "int32"
          },
          "open_blockers": {
            "type": "integer",
            "format": "int32"
          },
          "open_sub_issues": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "closed_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deleted_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deleted_by": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentityRef"
              }
            ]
          }
        }
      },
...
      "Label": {
        "required": [
          "name",
          "group",
          "description"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "group": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "NextAnswer": {
        "required": [
          "issue",
          "reasons"
        ],
        "type": "object",
        "properties": {
          "issue": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Issue"
              }
            ]
          },
          "reasons": {
            "$ref": "#/components/schemas/Reasons"
          }
        }
      },
      "NextPage": {
        "required": [
          "items",
          "total",
          "has_more",
          "reasons"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueSummary"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "has_more": {
            "type": "boolean"
          },
          "reasons": {
            "$ref": "#/components/schemas/Reasons"
          }
        }
      },
      "NextRequest": {
        "required": [
          "ready",
          "epic",
          "label",
          "repo",
          "limit"
        ],
        "type": "object",
        "properties": {
          "ready": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "epic": {
            "type": [
              "null",
              "string"
            ]
          },
          "label": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "repo": {
            "type": [
              "null",
              "string"
            ]
          },
          "limit": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
      "Priority": {
        "maximum": 4,
        "minimum": 0,
        "type": "integer"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "Project": {
        "required": [
          "key",
          "name",
          "triage_required",
          "review_required",
          "created_at",
          "updated_at"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "triage_required": {
            "type": "boolean"
          },
          "review_required": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Reasons": {
        "required": [
          "blocked",
          "waiting_for_answer",
          "in_progress",
          "in_review",
          "parked",
          "not_ready",
          "assigned_elsewhere"
        ],
        "type": "object",
        "properties": {
          "blocked": {
            "type": "integer",
            "format": "int32"
          },
          "waiting_for_answer": {
            "type": "integer",
            "format": "int32"
          },
          "in_progress": {
            "type": "integer",
            "format": "int32"
          },
          "in_review": {
            "type": "integer",
            "format": "int32"
          },
          "parked": {
            "type": "integer",
            "format": "int32"
          },
          "not_ready": {
            "type": "integer",
            "format": "int32"
          },
          "assigned_elsewhere": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}